+++ test-website/src/Pages/Test/TabManager/TabManager.tsx
@@ -4,10 +4,9 @@ import { PageProps } from 'Components/Routing/RouteProps'
 import { NavbarLink } from 'Components'
 import {
     Tab,
-    onChildReady,
     getTabFromLocation,
     TabManager,
-    onChildReady2
+    onChildReady
 } from '@interface-technologies/iti-react'
 import { TabClassesSection } from './TabClassesSection'
 import { TabContent } from './TabContent'
@@ -57,13 +56,13 @@ export class Page extends React.Component<PageProps, PageState> {
     }
 
     onChildReady = (args: Partial<Readiness>) => {
-        onChildReady2(x => this.setState(...x), args, () => {
+        onChildReady(x => this.setState(...x), args, () => {
             console.log('onChildReady callback - should only be called 3 times')
 
             if (this.isCurrentTabReady() && !this.props.ready) {
                 this.props.onReady({
                     title: 'Tab Test',
-                    activeNavbarLink: NavbarLink.Index,
+                    activeNavbarLink: NavbarLink.Index
                 })
             }
         })