+++ test-website/src/_Redux/Error/ErrorActions.ts
@@ -1,0 +1,6 @@
+﻿import { createStandardAction } from 'typesafe-actions'
+
+export const errorActions = {
+    onError: createStandardAction('ON_ERROR')<any>(),
+    clearError: createStandardAction('CLEAR_ERROR')()
+}